SpringORM

+++ spring-11-Data-TableCreation/src/main/java/com/cydeo/Repository/EmployeesRepository.java
@@ -3,5 +3,5 @@ package com.cydeo.Repository;
 import com.cydeo.entity.Employees;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeesRepository extends JpaRepository<Employees,Long> {
+public interface EmployeesRepository extends JpaRepository<Employees,Integer> {
 }
